Abstract ValueParserTestCase
