+++ tests/ValueParsers/YearMonthDayTimeParserTest.php
@@ -19,7 +19,7 @@ use ValueParsers\YearMonthDayTimeParser;
  * @license GPL-2.0-or-later
  * @author Thiemo Kreuz
  */
-class YearMonthDayTimeParserTest extends StringValueParserTest {
+class YearMonthDayTimeParserTest extends ValueParserTestCase {
 
 	/**
 	 * @see ValueParserTestBase::getInstance
@@ -179,7 +179,14 @@ class YearMonthDayTimeParserTest extends StringValueParserTest {
 			'2015-012-31',
 		);
 
-		$cases = parent::invalidInputProvider();
+		$cases = [
+			array( true ),
+			array( false ),
+			array( null ),
+			array( 4.2 ),
+			array( array() ),
+			array( 42 )
+		];
 
 		foreach ( $invalid as $value ) {
 			$cases[] = array( $value );
@@ -276,7 +283,7 @@ class YearMonthDayTimeParserTest extends StringValueParserTest {
 	 */
 	public function testInvalidOptions( array $options ) {
 		$parser = new YearMonthDayTimeParser( null, new ParserOptions( $options ) );
-		$this->setExpectedException( ParseException::class );
+		$this->expectException( ParseException::class );
 		$parser->parse( '2016-01-31' );
 	}
